Backend/requirements.txt
flask
openai
python-dotenv
sqlalchemy
psycopg2
pytz
